Refactor ModelsContent and related components to streamline session management and improve loading state handling
